refactor: extract DistributesIssuesForDisplayTrait from validators
